+++ src/Tests/WebApp.fs
@@ -191,8 +191,22 @@ let tests = testList "Web App Tests" [
         Expect.equal wa.ServerFarmId "[resourceId('my-asp-resource-group', 'Microsoft.Web/serverfarms', 'my-asp-name')]" ""
     }
 
+    test "Adds the Logging extension automatically for .NET Core apps" {
+        let wa = webApp { name "siteX" }
+        let extension = wa |> getResources |> getResource<SiteExtension> |> List.head
+        Expect.equal extension.Name.Value "Microsoft.AspNetCore.AzureAppServices.SiteExtension" "Wrong extension"
+
+        let wa = webApp { name "siteX"; runtime_stack Runtime.Java11 }
+        let extensions = wa |> getResources |> getResource<SiteExtension>
+        Expect.isEmpty extensions "Shouldn't be any extensions"
+
+        let wa = webApp { name "siteX"; automatic_logging_extension false }
+        let extensions = wa |> getResources |> getResource<SiteExtension>
+        Expect.isEmpty extensions "Shouldn't be any extensions"
+    }
+
     test "Handles add_extension correctly" {
-        let wa = webApp { name "siteX"; add_extension "extensionA"; }
+        let wa = webApp { name "siteX"; add_extension "extensionA"; runtime_stack Runtime.Java11 }
         let resources = wa |> getResources
         let sx = resources |> getResource<SiteExtension> |> List.head
         let r  = sx :> IArmResource
@@ -204,7 +218,7 @@ let tests = testList "Web App Tests" [
     }
 
     test "Handles multiple add_extension correctly" {
-        let wa = webApp { name "siteX"; add_extension "extensionA"; add_extension "extensionB"; add_extension "extensionB" }
+        let wa = webApp { name "siteX"; add_extension "extensionA"; add_extension "extensionB"; add_extension "extensionB"; runtime_stack Runtime.Java11 }
         let resources = wa |> getResources |> getResource<SiteExtension>
 
         let actual = List.sort resources